add self_adjust code

+++ Power_measurement_self_fitting_v1.0/Power_measurement_self_fitting_v1.0/Power_measurement_self_fitting_v1.0/src/Fitting_method.c
@@ -1,23 +1,23 @@
 #include "Fitting_method.h"
 
-float gather_xy = 0.0f;
-float gather_x = 0.0f;
-float gather_y = 0.0f;
-float gather_x2 = 0.0f;
-float gather_y2 = 0.0f;
-float gather_x_2 = 0.0f;
-float average_x = 0.0f;
-float average_y = 0.0f;
+double gather_xy = 0.0f;
+double gather_x = 0.0f;
+double gather_y = 0.0f;
+double gather_x2 = 0.0f;
+double gather_y2 = 0.0f;
+double gather_x_2 = 0.0f;
+double average_x = 0.0f;
+double average_y = 0.0f;
 
-float coe_a = 0.0f;
-float coe_b = 0.0f;
-float coe_r = 0.0f;
+double coe_a = 0.0f;
+double coe_b = 0.0f;
+double coe_r = 0.0f;
 
 
-void Get_average_x_y(float *array_x,float *array_y,uint32_t len)
+void Get_average_x_y(double *array_x,double *array_y,uint32_t len)
 {
 	uint32_t i = 0;
-	float sum_x=0.0f,sum_y=0.0f;
+	double sum_x=0.0f,sum_y=0.0f;
 	
 	for(i=0;i<len;i++)
 	{
@@ -29,7 +29,7 @@ void Get_average_x_y(float *array_x,float *array_y,uint32_t len)
 	average_y = sum_y/len;
 }
 
-void Get_gather_all(float *array_x,float *array_y,uint32_t len)
+void Get_gather_all(double *array_x,double *array_y,uint32_t len)
 {
 	uint32_t i = 0;
 	
@@ -45,12 +45,12 @@ void Get_gather_all(float *array_x,float *array_y,uint32_t len)
 	gather_x_2 = gather_x*gather_x;
 }
 
-float Get_coe_r(float *array_x,float *array_y,uint32_t len)
+double Get_coe_r(double *array_x,double *array_y,uint32_t len)
 {
 	uint32_t i = 0;
-	float gather_x_X=0,gather_y_Y=0;
-	float gather_x_X_2=0,gather_y_Y_2=0;
-	float ret = 0;
+	double gather_x_X=0,gather_y_Y=0;
+	double gather_x_X_2=0,gather_y_Y_2=0;
+	double ret = 0;
 
 	for(i=0;i<len;i++)
 	{
@@ -70,7 +70,7 @@ float Get_coe_r(float *array_x,float *array_y,uint32_t len)
 	return ret;
 }
 
-void Get_coe_a_b_r(float *array_x,float *array_y,uint32_t len,float *coea,float *coeb,float *coer)
+void Get_coe_a_b_r(double *array_x,double *array_y,uint32_t len,double *coea,double *coeb,double *coer)
 {
 	Get_gather_all(array_x,array_y,len);
