+++ Power_measurement_self_fitting_v1.0/Power_measurement_self_fitting_v1.0/Power_measurement_self_fitting_v1.0/src/Fitting_method.h
@@ -3,9 +3,9 @@
 
 #include "stm32f10x.h"
 
-void Get_average_x_y(float *array_x,float *array_y,uint32_t len);
-void Get_gather_all(float *array_x,float *array_y,uint32_t len);
-float Get_coe_r(float *array_x,float *array_y,uint32_t len);
-void Get_coe_a_b_r(float *array_x,float *array_y,uint32_t len,float *coea,float *coeb,float *coer);
+void Get_average_x_y(double *array_x,double *array_y,uint32_t len);
+void Get_gather_all(double *array_x,double *array_y,uint32_t len);
+double Get_coe_r(double *array_x,double *array_y,uint32_t len);
+void Get_coe_a_b_r(double *array_x,double *array_y,uint32_t len,double *coea,double *coeb,double *coer);
 
 #endif
